requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-downloadview==1.9
django-filetransfers==0.1.0
django-filter==2.1.0
djangorestframework==3.9.2
idna==2.8
Markdown==3.1
Pillow==6.0.0
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
